3.0.0

+++ src/main/java/net/fellter/vanillalayerplus/custom_blocks/SlimeLayerBlock.java
@@ -15,7 +15,7 @@ public class SlimeLayerBlock extends LayerBlock {
 		super(settings);
 	}
 
-	public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, float fallDistance) {
+	public void onLandedUpon(World world, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
 		if (!entity.bypassesLandingEffects()) {
 			entity.handleFallDamage(fallDistance, 0.0F, world.getDamageSources().fall());
 		}
